aider: Added loopback address 127.0.0.1 to the `privateIPBlocks`.
# Aider chat conversation:

USER: why is the local loop-back ip 127.0.0.1 not included in the p`privateIPBlocks`?

ASSISTANT: The local loopback IP address 127.0.0.1 is not included in the `privateIPBlocks` because it is not typically considered part of the private IP address space.

Private IP addresses are defined by the Internet Assigned Numbers Authority (IANA) and are reserved for use within private networks. These addresses are not routable on the public internet, which means they cannot be reached from outside the network they are used in. The ranges you see in the `privateIPBlocks` are the ones defined by IANA as private IP addresses.

The IP address 127.0.0.1 is a loopback address, which means it is used by a host to send a network request to itself. It is used for testing and network diagnostics, and is not part of the private IP address space.

However, if you want to include the loopback address in the `privateIPBlocks`, you can add it to the list. Here's how you can do it:

```go
cmd/geoip.go
<<<<<<< HEAD
privateIPBlocks := []net.IPNet{
	{IP: net.IPv4(10, 0, 0, 0), Mask: net.CIDRMask(8, 32)},
	{IP: net.IPv4(172, 16, 0, 0), Mask: net.CIDRMask(12, 32)},
	{IP: net.IPv4(192, 168, 0, 0), Mask: net.CIDRMask(16, 32)},
}
=======
privateIPBlocks := []net.IPNet{
	{IP: net.IPv4(10, 0, 0, 0), Mask: net.CIDRMask(8, 32)},
	{IP: net.IPv4(172, 16, 0, 0), Mask: net.CIDRMask(12, 32)},
	{IP: net.IPv4(192, 168, 0, 0), Mask: net.CIDRMask(16, 32)},
	{IP: net.IPv4(127, 0, 0, 1), Mask: net.CIDRMask(8, 32)},
}
>>>>>>> updated
```

This will add the loopback address to the list of private IP blocks, and the `isPrivateIP` function will return `true` if the IP address is 127.0.0.1.