aider: Added timing code to measure the run time for a read of a logfile.
# Aider chat conversation:

USER: Please add timing code to the central run function,that will measure the run time for a read of a logfile. The output should be human readable in the form of hh:mm:ss - the percission of 1 second is enough

ASSISTANT: To add timing code to the central run function, we need to modify the `cmd/root.go` file. We will use the `time` package in Go to measure the execution time of the read operation.

Here are the steps we will follow:

1. At the start of the `Run` function, we will record the current time using `time.Now()`.
2. At the end of the `Run` function, we will again record the current time.
3. We will subtract the start time from the end time to get the duration of the read operation.
4. We will print the duration in a human-readable format.

Here is the *edit block* for the changes:

```go
cmd/root.go
<<<<<<< HEAD
	Run: func(cmd *cobra.Command, args []string) {
=======
	Run: func(cmd *cobra.Command, args []string) {
		startTime := time.Now()
>>>>>>> updated

...

```go
cmd/root.go
<<<<<<< HEAD
		fmt.Printf("Count of entries: %d\n", len(LogEntries))
	},
=======
		endTime := time.Now()
		duration := endTime.Sub(startTime)
		fmt.Printf("Count of entries: %d\n", len(LogEntries))
		fmt.Printf("Time taken to read the logfile: %s\n", duration)
	},
>>>>>>> updated
```

Please apply these changes to your code.