aider: Ensure that if the user enters a partial datetime for the End variable, the value written to the variable is the latest datetime of that day.
# Aider chat conversation:

USER:
Can we ensure that if the user enters a parial datetime, say only year and day for the End var that:
- the value written to the var is actually the latest datetime of that day? As in 2023-12-01 would set the End date to midnight, on that day

ASSISTANT: Yes, we can achieve this by modifying the `setCmd` function in the `set.go` file. We will need to check if the parsed date has a time component. If it doesn't, we will set the time to the end of the day (23:59:59).

Here are the steps:

1. Parse the date string using `dateparse.ParseAny(args[1])`.
2. Check if the parsed date's time component is zero (i.e., it's the start of the day).
3. If it is, add 23 hours, 59 minutes, and 59 seconds to it to make it the end of the day.
4. Format the updated date and assign it to the `End` variable.

Here is the *edit block* for the changes:

```go
cmd/set.go
<<<<<<< HEAD
			case "end":
				if t, err := dateparse.ParseAny(args[1]); err == nil {
					End = t.Format(time.RFC3339)
					fmt.Printf("Updated 'end' to: %s\n", End)
				} else {
					fmt.Printf("Invalid date format for end: %s\n", args[1])
				}
=======
			case "end":
				if t, err := dateparse.ParseAny(args[1]); err == nil {
					if t.Hour() == 0 && t.Minute() == 0 && t.Second() == 0 {
						t = t.Add(time.Hour*23 + time.Minute*59 + time.Second*59)
					}
					End = t.Format(time.RFC3339)
					fmt.Printf("Updated 'end' to: %s\n", End)
				} else {
					fmt.Printf("Invalid date format for end: %s\n", args[1])
				}
>>>>>>> updated
```

This change will ensure that if a user enters a partial date for the `End` variable, the value will be set to the end of that day.