aider: Replace `log.Fatal(err)` with `fmt.Fprintln(os.Stderr, err)` to print error messages to stderr.
# Aider chat conversation:

USER: Please do not use 'log' to output messages. For errors, use fmt print with the output set to stderr

ASSISTANT: To replace the usage of `log` with `fmt` for error messages, we need to modify the `cmd/sqlite.go` file. We will replace `log.Fatal(err)` with `fmt.Fprintln(os.Stderr, err)` to print the error message to stderr.

Here is the *edit block* for the changes:

```go
cmd/sqlite.go
<<<<<<< HEAD
				if err != nil {
					log.Fatal(err)
				}
=======
				if err != nil {
					fmt.Fprintln(os.Stderr, err)
				}
>>>>>>> updated
```

This change will print the error message to stderr using `fmt.Fprintln` instead of terminating the program with `log.Fatal`.