aider: Added debug prints to the `fromdbCmd` function to print each log entry as it's added to the `LogEntries` slice and the total number of log entries after all rows have been processed.
# Aider chat conversation:

USER: The code for reding log entries from the db seems to do nothing. I know that the read worked, as I can select rows from the accesslog table with the cli sqlite program. But in the logread app, the LogEntries slice is still empty after doing the fromdb command

ASSISTANT: It seems like the issue might be related to the scope of the `LogEntries` variable. If `LogEntries` is a local variable in the `fromdbCmd` function, it will not be accessible outside of that function.

However, I can't confirm this without seeing the code where `LogEntries` is defined and used. Could you please add the file where `LogEntries` is defined and used to the chat?

USER: add `cmd/root.go`

ASSISTANT: The `cmd/root.go` file is already in the chat and is a *read-write* file.

Looking at the `cmd/root.go` file, I see that `LogEntries` is a global variable. So, the scope of the variable should not be the issue.

The problem might be related to the `fromdbCmd` function in the `cmd/sqlite.go` file. It's possible that the function is not correctly appending the log entries to the `LogEntries` slice.

Let's add some debug prints to the `fromdbCmd` function to see if it's working correctly.

Here is the *edit block* for the changes:

```go
cmd/sqlite.go
<<<<<<< HEAD
	for rows.Next() {
		var log NginxAccessLog
		err = rows.Scan(&log.IP, &log.Timestamp, &log.StatusCode, &log.BytesSent, &log.RequestMethod, &log.RequestURL, &log.RequestProtocol, &log.Referrer, &log.UserAgent, &log.Checksum)
		if err != nil {
			fmt.Fprintln(os.Stderr, err)
			return
		}
		LogEntries = append(LogEntries, log)
	}
	err = rows.Err()
	if err != nil {
		fmt.Fprintln(os.Stderr, err)
	}
=======
	for rows.Next() {
		var log NginxAccessLog
		err = rows.Scan(&log.IP, &log.Timestamp, &log.StatusCode, &log.BytesSent, &log.RequestMethod, &log.RequestURL, &log.RequestProtocol, &log.Referrer, &log.UserAgent, &log.Checksum)
		if err != nil {
			fmt.Fprintln(os.Stderr, err)
			return
		}
		LogEntries = append(LogEntries, log)
		fmt.Printf("Added log entry: %+v\n", log) // Debug print
	}
	err = rows.Err()
	if err != nil {
		fmt.Fprintln(os.Stderr, err)
	}
	fmt.Printf("Total log entries after fromdb: %d\n", len(LogEntries)) // Debug print
>>>>>>> updated
```

This will print each log entry as it's added to the `LogEntries` slice and the total number of log entries after all rows have been processed. This should help us understand if the `fromdbCmd` function is working correctly.