aider: Updated `fromdbCmd` command in `cmd/sqlite.go` to parse the timestamp string into a `time.Time` object.
# Aider chat conversation:

ASSISTANT: To convert the string to a `time.Time` object, we need to use the `time.Parse` function. We need to know the format of the timestamp string to correctly parse it.

Assuming the timestamp is in the format "2006-01-02T15:04:05Z07:00", we can modify the `fromdbCmd` command in the `cmd/sqlite.go` file to parse the timestamp string into a `time.Time` object.

Here is the change:

```go
cmd/sqlite.go
<<<<<<< HEAD
			err = rows.Scan(&log.IP, &log.Timestamp, &log.StatusCode, &log.BytesSent, &log.RequestMethod, &log.RequestURL, &log.RequestProtocol, &log.Referrer, &log.UserAgent, &log.Checksum)
=======
			var timestampStr string
			err = rows.Scan(&log.IP, &timestampStr, &log.StatusCode, &log.BytesSent, &log.RequestMethod, &log.RequestURL, &log.RequestProtocol, &log.Referrer, &log.UserAgent, &log.Checksum)
			if err != nil {
				fmt.Fprintln(os.Stderr, err)
				return
			}
			log.Timestamp, err = time.Parse(time.RFC3339, timestampStr)
			if err != nil {
				fmt.Fprintln(os.Stderr, err)
				return
			}
>>>>>>> updated
```

This change will replace the line that scans the row data into the `log` object with a block of code that scans the timestamp into a separate string variable, parses that string into a `time.Time` object, and assigns the result to `log.Timestamp`.

Please note that this change assumes that the `Timestamp` field of the `NginxAccessLog` struct is of type `time.Time`. If it's not, you'll need to change the type of this field to `time.Time` in the `cmd/root.go` file. If you need to make this change, please add the `cmd/root.go` file to the chat.